backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
pillow==10.4.0
numpy==1.26.4
scipy==1.13.0
opencv-python==4.10.0.84
pyexiftool==0.5.6
torch==2.3.0
torchvision==0.18.0
transformers==4.43.0
uniface==1.6.0
huggingface-hub==0.24.0
python-magic==0.4.27
